Added a property that turn on/off  gzip compression of css and js files.
If suppose, gzip compression is turned on server for all the css files
and granule combined css file can not be omitted from this compression.
gzip compression happens two times, and the combined css / js files may
not work as expected.  So, there must be an option in granule to turn
off gzip compression. a property called gzip.output is added to
configuration. By default this value is set to true. If granule gzip
compression is not required, then this parameter has to be set to false.